backend/pyproject.toml
[tool.poetry]
name = "ai-vc-copilot"
version = "0.1.0"
description = "AI VC Co-Pilot — Agentic venture consultancy platform"
authors = ["AI VC Co-Pilot Team"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
fastapi = {extras = ["standard"], version = "^0.115.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.0"}
asyncpg = "^0.31.0"
alembic = "^1.14.0"
pydantic = {extras = ["email"], version = "^2.0.0"}
pydantic-settings = "^2.0.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
bcrypt = "^4.2.0"
boto3 = "^1.35.0"
celery = {extras = ["redis"], version = "^5.4.0"}
redis = "^5.0.0"
openai = "^1.50.0"
anthropic = "^0.40.0"
psycopg2-binary = "^2.9.0"
structlog = "^24.0.0"
pypdf = "^5.0.0"
python-pptx = "^1.0.0"
python-docx = "^1.1.0"
openpyxl = "^3.1.0"
pgvector = "^0.3.0"
python-multipart = "^0.0.12"
uvicorn = "^0.32.0"
deepdiff = "^8.0.0"
jinja2 = "^3.1.0"
weasyprint = "^62.0"
pyyaml = "^6.0"
slowapi = "^0.1.9"
prometheus-fastapi-instrumentator = "^7.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.24.0"
httpx = "^0.28.0"
aiosqlite = "^0.21.0"
ruff = "^0.8.0"
mypy = "^1.13.0"
types-python-jose = "*"
types-boto3 = "*"
types-openpyxl = "*"
types-redis = "*"
types-pyyaml = "^6.0.12.20250915"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "SIM"]
ignore = ["A003", "B008"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
